+++ workspace/solutions/kotlin/lib/src/test/kotlin/utils/Test_.kt
@@ -18,9 +18,11 @@
 package utils
 
 // REF: https://kotlinlang.org/api/latest/kotlin.test/kotlin.test/
+import kotlinx.serialization.json.buildJsonArray
 import kotlin.reflect.KCallable
 import kotlin.reflect.KClass
 import kotlin.reflect.KVisibility
+import kotlin.reflect.full.createInstance
 import kotlin.test.*
 
 public class Test_(
@@ -29,6 +31,8 @@ public class Test_(
 ) {
     val packageName: String
     val Solutions: Array<out KClass<*>>
+
+    lateinit var examples: List<Example>
 
     init {
         this.packageName = SolutionsTest
@@ -43,7 +47,7 @@ public class Test_(
          * @param Solution 题解类
          * @return 题解的入口方法
          */
-        private final fun getMethod(Solution: KClass<*>): KCallable<*> {
+        final fun getMethod(Solution: KClass<*>): KCallable<*> {
             // REF: https://kotlinlang.org/api/latest/jvm/stdlib/kotlin.reflect/-k-class/
             val methods = Solution
                 .members
@@ -68,10 +72,11 @@ public class Test_(
 
         /**
          * 通过反射获取方法的参数类型名列表
+         * 已排除第一个参数 (方法绑定的类)
          * @param method 方法
          * @return 方法的参数类型名列表
          */
-        private final fun getMethodParameterTypeNames(method: KCallable<*>): List<String> = method.parameters.map { it.type.toString() }
+        private final fun getMethodParameterTypeNames(method: KCallable<*>): List<String> = method.parameters.drop(1).map { it.type.toString() }
 
         /**
          * 通过反射获取方法的返回类型名
@@ -79,6 +84,49 @@ public class Test_(
          * @return 方法的返回类型名
          */
         private final fun getMethodReturnTypeName(method: KCallable<*>): String = method.returnType.toString()
+
+        /**
+         * 使用一个示例测试题解
+         * @param solution 题解实例
+         * @param example 示例
+         * @param solutionIndex 题解索引
+         * @param exampleIndex 示例索引
+         */
+        private final fun testExample(
+            solution: Any,
+            example: Example,
+            solutionIndex: Int,
+            exampleIndex: Int,
+        ) {
+            val method = Test_.getMethod(solution::class)
+            val result = method.call(solution, *example.input)
+            assertNotNull(result, "Result is null")
+
+            val result_type_name = result::class.qualifiedName
+            val output_type_name = example.output::class.qualifiedName
+            assertEquals(
+                result_type_name,
+                output_type_name,
+                "Solution result type <$result_type_name> != expected type <$output_type_name>",
+            )
+
+            val parameter_type_names = Test_.getMethodParameterTypeNames(method)
+            val return_type_name = Test_.getMethodReturnTypeName(method)
+
+            val input_json = buildJsonArray {
+                example.input.withIndex().forEach { (index, value) ->
+                    add(Example.any2json(value, parameter_type_names[index]))
+                }
+            }
+            val result_json = Example.any2json(result, return_type_name)
+            val output_json = Example.any2json(example.output, return_type_name)
+
+            assertEquals(
+                result_json,
+                output_json,
+                "\nsolution: $solutionIndex\nexample:  $exampleIndex\ninput:    $input_json\nresult:   $result_json\nexpected: $output_json\n",
+            )
+        }
     }
 
     public final fun run() {
@@ -91,8 +139,17 @@ public class Test_(
     private final fun test() {
         this.testSolutionsTypeDefinition()
 
-        // TODO: 运行题解
-        // TODO: 校验题解结果
+        for ((solutionIndex, Solution) in this.Solutions.withIndex()) {
+            val solution = Solution.createInstance()
+            for ((exampleIndex, example) in this.examples.withIndex()) {
+                Test_.testExample(
+                    solution,
+                    example,
+                    solutionIndex,
+                    exampleIndex,
+                )
+            }
+        }
     }
 
     /**
@@ -133,8 +190,16 @@ public class Test_(
             )
         }
 
-        // TODO: 加载题解测试示例
-        // TODO: 构造题解测试示例对象
-        assertTrue(false, "$method_return_type_name")
+        val examples_json = Example.readExamplesFile(this.packageName)
+        assertNotNull(
+            examples_json,
+            "No examples",
+        )
+
+        this.examples = Example.fromJson(
+            examples_json,
+            method_parameter_type_names,
+            method_return_type_name,
+        )
     }
 }